Fix in breadcrumb generation for non-technology pages.
The homepage as well as the privacy policy, which are of type HOMEPAGE and TEXT respectively, are derived from a given technology. Therefore the breadcrumb path has to be built differently. The privacy policy was also updated as part of this fix; removing the HSTS preload which is not applicable on Github pages as github.io is a public domain.