RK script now handling post-name typing.
Certain types for function parameters, such as arrays typically, often require the prototype to have a certain part before the name of the parameter, and some after it. In C, it may give my_function(int a[]) for instance. This is now supported by having a special character in the type name; a tabulation (\t) which acts as a delimited between left-side type and right-side type content.