Extraneous multiplication removed from MPI_IN_PLACE C example. #317
The C-version of the examples from MPI_IN_PLACE documentation was containing an extra multiplication. Given that the MPI process on which it was happening had rank 0 in that configuration, it had no effect but remained an unwanted additional value change that had to be removed. However, this could lead to the design of another exercise that would focuss on the fact that, with MPI_IN_PLACE, the buffer passed by the root MPI process is left untouched for the element(s) assigned to the root MPI process.